[fs] fix bugs in glob and list in our GSFS (#13434)
CHANGELOG: In QoB, Hail's file systems now correctly list all files in a
directory, not just the first 1000. This could manifest in an
import_table or import_vcf which used a glob expression. In such a case,
only the first 1000 files would have been included in the resulting
Table or MatrixTable.

I also moved two GSFS-only tests into the FSSuite. There should be very
few tests that are cloud-specific.