[qob] cleanup QoB tests without erroring
In main, we assumed that if `self._batch is not None` then there is a Batch that
needs cancelling, but that is no longer true. Now, we must track whether we have
called `submit` or not.

I made a few other minor cleanups of `ServiceBackend` while I was in there. There
is no longer any need to have a `None` batch b/c there is no distinction between
a builder and a batch now.