[qob][batch] do not list all jobs on failure (plus: types!) (#13500)
CHANGELOG: In Query-on-Batch, the client-side Python code will not try
to list every job when a QoB batch fails. This could take hours for
long-running pipelines or pipelines with many partitions.

I also added API types to the list jobs end point because I have to go
hunting for this every time anyway. Seems better to have this
information at our digital fingertips.