[test-dataproc] make test-dataproc-37 and -38 not race
Consider, for example, this deploy: https://ci.hail.is/batches/7956812. `test-dataproc-37` succeeded
but `test-dataproc-38` failed (it timed out b/c the master failed to come online).

You can see the error logs for the cluster here: https://cloudlogging.app.goo.gl/t1ux8oqy11Ba2dih7

It states a certain file either did not exist or we did not have permission to access it.

[`test_dataproc-37`](https://batch.hail.is/batches/7956812/jobs/193) and [`test_dataproc-38`](https://batch.hail.is/batches/7956812/jobs/194)
started around the same time and both uploaded four files into:

    gs://hail-30-day/hailctl/dataproc/ci_test_dataproc/0.2.121-7343e9c368dc/

And then set it to public read/write. The public read/write means that permissions are not the issue.

Instead, the issue is that there must be some sort of race condition in GCS which means that if you
"patch" (aka overwrite) an existing file, it is possible that a concurrent reader will see the file
as not existing.