[dataproc][vep] mitigate Docker bug
CHANGELOG: Mitigate #12936 in which VEP Dataproc clusters fail to start. The root cause is complex. Docker has a bug which prevents it from cleanly starting if it is *re* installed. Whatever Google is doing in Dataproc to configure their Docker "component" appears to trigger this bug.

See for details: https://github.com/hail-is/hail/issues/12936#issuecomment-1709120751

The basic fix is to sleep to allow the system to coalesce a bit and then to restart Docker.