[fs] consistent use of the term FileListEntry (#13594)
In our asyncio FSes, we use the term FileListEntry for the package of
information you receive from a cloud object store list operation. We
reserve the term FileStatus for the metadata you get about a blob
without a list operation. A major difference is that FLEs know if they
refer to files or folders. In contrast, a FileStatus knows that it is a
file *but does not know if a directory with the same name exists*.