[batch] Fix batch local mode in jupyter notebooks (#13614)
The `LocalBackend` recently was changed to include a call to
`async_to_blocking` but does not apply `nest_asyncio`, so using the
local backend in a jupyter notebook breaks. I lifted the application of
`nest_asyncio` from the `ServiceBackend` to the super class. Should we
add a test that runs in its own python session to make sure that this is
caught?