[auth|batch|aiogoogle] Favor latent credentials over specifying /gsa-key/key.json in application code (#13596)
This is mostly cleanup reducing the amount of the codebase that depends
on GSA key files (which are not available in terra or a future keyless
hail-vdc). The core bit is instead of threading
`credentials_file="/gsa-key/key.json` through the batch and auth
codebase we set `GOOGLE_APPLICATION_CREDENTIALS` in their deployments. I
can't think of a scenario where `auth` or `batch` would need to use
multiple identities so better to remove their ability to do so and
always use the default credentials.

I also did a bit of tidying up, using `$GOOGLE_APPLICATION_CREDENTIALS`
instead of the hard-coded path and removing the credentials endpoints on
the batch-driver which have been unused by workers for many months now.