[qob] don’t use unbounded threads reading results (#13634)
CHANGELOG: fix a bug which caused jobs with a large number of partitions
(close to 100k) to run out of memory on the driver after all partitions
finish.

Cap number of threads allocated when reading results of a cda on the
driver.

~Still waiting on a test to confirm this fixes the bad scaling with
number of partitions.~

We were reading cda results with a threadpool containing one thread per
partition, but kept all but 1000 threads blocked. This changes to use a
threadpool of 1000 threads.

We still use a surprising amount of memory when reading results, but
this change will likely allow jobs with ~100k partitions to complete.

fixes #13533