[query] non-shadow-jar for pytest, doctest, and install-editable
Apparently, creating a zip file of all our dependencies and the Hail code takes ~30s. This change
skips the shadow JAR for local development. We still need a shadow JAR to produce a shareable wheel
file; however, when doing local development, we can simply tell Py4J (and the JVM) where to find our
dependencies class files.

Also notice that I made install-editable non-PHONY. It need not be PHONY as long as we can reliably
determine if `hail/python` is the pip-installed version. To do so, we simply check if the
__init__.py at the root of the pip package is newer than when we last installed. If its newer, then
either:
1. We edited __init__.py, or
2. The pip location of Hail has changed since we last ran install-editable

I also deleted eggs because nobody uses eggs.