[query] Call-Cache `CollectDistributedArray` (rfc-0000) (#12954)
Adds the ability to rerun/retry queries from the nearest
`CollectDistributedArray` (`CDA`) `IR` site.

Computes a "Semantic Hash" of the top-level `IR` which is used to
generate a key for the various constituent `CDA` calls in a query. The
implementation for CDA, `BackendUtils.collectDArray`, uses that key to
look into an the execution cache for the results of each partition for
that call and uses/updates the cache with successful partition
computations.

The nature of the staged- lower and execute model means we don't know
how many `CDA` calls that will be generated ahead of time. Thus we treat
the "Semantic Hash" in a similar way to an RNG state variable and
generate a key from the Semantic Hash every time every time we encounter
a `CDA`.

The execution cache is implemented on-top of a local or remote
filesystem (configurable via the `HAIL_CACHE_DIR` environment variable).
This defaults to `{tmpdir}/hail/{pip-version}`.