[query/ggplot] runs plotting notebook setup on import (#13610)
CHANGELOG: Additional setup is no longer required when using `hail.plot`
or `hail.ggplot` in a Jupyter notebook (calling
`bokeh.io.output_notebook` or `hail.plot.output_notebook` and/or setting
`plotly.io.renderers.default = 'iframe'` is no longer necessary).

I double-checked locally that `import hail as hl` and `from hail.ggplot
import ggplot` both cause the added code to run, and that running them
in a notebook does the setup we expect it to, but running it in iPython
or the Python interpreter does not.