[jvm-fs] rename listStatus to listDirectory
The "status" in `listStatus` referred to `FileStatus`, a Hadoop abstraction of a Linux
`stat` object. In my last PR, I split that concept into `FileStatus` and `FileListEntry`
where the former is only for files and the latter is for the contents of a directory
(in the latter case, cloud providers usually definitively tell you if a path is an object,
an object and a prefix [aka directory], or a prefix).

This method is really about listing (a la `ls`) directories, so I have changed its
name to reflect that.

I specifically choose not to name it `ls` or `list` because I do not want it confused
with `ls` because `ls` typically supports "listing" a specific file. This method
is only for listing the contents of a directory.