[jvm-fs] simplify use of URLs
This change enforces that `String` accepting methods just convert to `URL` and do nothing more.
All the logic is now in `URL` accepting methods. I think this is a bit easier to
understand. It also avoids the `val url = XXXX.parseUrl(filename)` at the start of every
method. I had to remove the type parameter because I would have needed a kind of
existential types that I could not easily encode in Scala.