[auth] Opt out of CSRF check for Envoy auth endpoints (#13646)
The comment in the code has most of the context. For auth checks, Envoy
uses the same HTTP method as the original request, which confuses our
aiohttp middleware which checks CSRF tokens on every `POST`. This broke
Grafana because the grafana front-end POSTs queries to the backend, and
those end up failing the auth check. We just don't need to check for
CSRF in auth checks. This should be the only occurrence in our codebase
where a safe endpoint can be reached with an unsafe HTTP method.