[query/ggplot] makes plotting setup lazy (#13648)
https://github.com/hail-is/hail/pull/13610 added the setup for bokeh and
plotly into `hailtop/__init__.py`, which causes `IPython` to be imported
when invoking `hailctl`, adding unnecessary startup time to it (observed
between 0.210s and 0.400s). This change moves the setup into the
`__init__.py` files for the plotting modules, with the resulting state
stored in a global variable in `hailtop/__init__.py`. The plotting setup
is still run upon invoking `import hail`, as this imports all submodules
of hail as well, but the `IPython` import no longer happens when
invoking `hailctl`.