Add Support for G2 Machines to Batch (#13430)
Currently, jobs in hail batch can only be run on n1 machines but with
the rise of deep learning in bioinformatics, the ability to run jobs on
g2 machines, as well as other GPU supported machines, is an important
and exciting addition to hail batch. This PR highlights the steps needed
to add new machine types into hail batch and could be used as a template
for further development support.

The changes in this PR can broadly be divided into additions to the job
crun container and insertion of g2 resources (CPU, RAM, L4 Accelerator)
into the resources table for billing. This PR uses the NVIDIA Container
Toolkit, which allows the creation of GPU accelerated containers. This
toolkit is integrated with docker via the parameters —runtime=nvidia and
the specification of GPUs is made through —gpus all. The toolkit is
installed in the batch worker VM startup script and the corresponding
docker parameters are configured if the machine type is g2, so there is
no change to the docker configuration for n1 machines. For the toolkit
to work there is a nvidia hook that needs to be injected into the crun
config. These modifications are also done based on machine type. On the
billing side, the existing pricing setup was expanded to include g2
machines. The g2 instance cores and RAM are inserted into the database,
and the SKUs are hard coded. For future machine type incorporation or
updates,
[https://cloud.google.com/skus/?currency=USD&filter=](https://cloud.google.com/skus/?currency=USD&filter=)
may serve as a useful resource to identify relevant SKU ids. A new
resource type was also added for the accelerator, including preemptible
and non-preemtible. Finally, g2 machines mount the worker data disk
under the name nvme0n2 so the code is updated to reflect this. Future
work may want to investigate a way to automatically detect what the
proper disk name is or make the disk naming logic more robust.

---------

Co-authored-by: Sophie Parsa <parsa@wm9c6-e4e.broadinstitute.org>
Co-authored-by: Daniel Goldstein <danielgold95@gmail.com>