[batch] Account for JVM containers when allocating network namespaces (#13678)
Creating network namespaces can often take hundreds of milliseconds (and
sometimes seconds with `iptables` contention), so Batch takes this off
the job hot path by pre-allocating namespaces. All job namespaces are
configured identically and there is a fixed number of "slots" on any
batch worker (`CORES * 4`), so pre-allocation and asynchronous recycling
of namespaces is fairly straight-forward so long as we never attempt to
run more containers on a worker than the number of slots (which the
scheduling system should prohibit).

However, since we started running long-lived JVM containers (#11397),
the number of containers running on a given worker can easily be
*greater* than `N_SLOTS`. On a 16-core machine, we create 30 JVMs that
sit idle waiting for JVMJobs all the while occupying a precious network
namespace.

I thought for the longest time that #13402 was a race condition so was
trying to trigger it through a barrage of quick jobs. Turns out all it
took was running >34 long-running jobs on a single 16-core worker. In a
dev deploy of `main`, running a batch with 35 quarter-core `sleep 150`
jobs fails with a single job timing out waiting for a network. On this
branch, I am able to run the same 35 job batch as well as a batch with
64 quarter-core jobs.

Unfortunately, we don't have a great way to test "run all these jobs at
once on the same worker".

Resolves #13402