[hailtop] replace humanize.naturaldelta (#13673)
Specifically for time, humanize does not respect significant figures.
Consider this:

```
In [7]: humanize.naturaldelta(120-1)
Out[7]: 'a minute'

In [8]: humanize.naturaldelta(120)
Out[8]: '2 minutes'
```

While the difference between one minute and two minutes is minor, the
difference between one hour and two hours is rather large:

```
In [6]: humanize.naturaldelta(2*60*60-1)
Out[6]: 'an hour'

In [7]: humanize.naturaldelta(2*60*60)
Out[7]: '2 hours'
```

This leads to user confusion because the cost of a two hour job seems to
be substantially more than "an hour" job.