[qob] prevent race to write to log file
This finally block is currently located after `gather.take()` but *before* we cancel (aka shutdown)
all threads. As a result, it is possible for us to shudown the logging (thus flushing, closing,
and destorying old appenders) then restart logging (thus opening the file in overwrite mode) and
blow away whatever was there.

I have verified in my namespace across 10s of thosuands of JVM Jobs that this never blows away the
log.