[query] fix backoff code (#13713)
CHANGELOG: Fixes #13704, in which Hail could encounter an
IllegalArgumentException if there are too many transient errors.

I need to do the multiplication in 64-bits so that it does not wrap
around to a large negative value. Then I can use `math.min` with the
maxDelayMs to get us back into 32-bits.

I'm just pushing through a bunch of bugs to get Wenhan unblocked today.