[auth] Reuse auth decorators without making additional requests (#13672)
Currently the auth service uses its own authentication decorators
instead of using those that the other services do (because those make a
request to auth which feels a bit circular), but now we have two sets of
decorators which we want to mostly keep in step aside from where they
get their `UserData` from. This PR creates a small abstraction with an
`Authenticator` base class and a `AuthServiceAuthenticator` subclass
that extends it and fetches userdata from the auth service. The auth
service instead uses a `LocalAuthenticator` which instead of making an
rpc directly fetches the userdata from its database.

(I would recommend using the split diff)