[qob] include version in batch attributes (#13731)
This would have been useful when I was rapidly iterating on different
versions. It saves me from clicking into a job to figure out which JAR
was used and then working backward from git SHA to Hail version.

I also noticed we had this deadcode which was supposed to set the name.
I moved that into the constructor so that it has the intended effect.