Gateway return 503 on failed auth check (#13735)
Envoy by default responds with a 403 if it fails to make an
authorization check. We should treat the failure to contact `auth` as a
transient error, so that clients can know to retry whatever request they
were trying to make instead of failing. In particular, in dev namespaces
the current behavior can trigger a QoB client to cancel an ongoing
pipeline while polling for completion.
[status_on_error](https://www.envoyproxy.io/docs/envoy/latest/api-v3/extensions/filters/http/ext_authz/v3/ext_authz.proto#extensions-filters-http-ext-authz-v3-extauthz)
allows us to configure that default behavior to return a 503
unavailable.

I deployed this in Azure by running `make -C gateway deploy
NAMESPACE=default`. I poked around to see that I could access my dev
namespace and production pages through the browser but did not otherwise
try to prove that this failure mode no longer exists.