[query] Deprecate `collect_all` in `hail.plots.*` (use `n_divisions` instead) (#13738)
Fixes #13696.
`n_divisions` is a downsampling factor. It conflicts with `collect_all`
when `collect_all=True` is specified.
Deprecate `collect_all` and issue warning when users supply a non-`None`
`collect_all` argument.
Use `n_divisions=None` to not downsample (implies `collect_all=True`).
Throw when `collect_all=True` and `n_divisions` is not `None` - this is
a conflict.