[humanizex] Run humanize tests and fix microseconds bug (#13743)
This is two independent commits but I put them in one PR, happy to split
if you want.

1. These tests weren't being run in CI I think because of the filename
not starting with `test_*`. The microseconds test currently fails on
main. I believe this is because the fallback case at the end uses the
units of the input and not the smallest possible units (on main you get
naturaldelta(0.000001) == 0 seconds not 1 microsecond)

2. Changed second(s) -> s, millisecond(s) -> ms, microsecond(s) -> µs