[query] do not store function return values in variables
I do not know why we do this, but we return a `Code` here, not a `Value`, so it seems silly to store
in a variable. In practice, the LIR is littered with invocations that store their value in a local
which is then read and stored in a different local beause we `memoize` the function call.