[query] fix ndarray concat with size 0 dims (#13755)
ndarray concat was broken when the first input has size 0 along the
concat axis. For example
```
In [3]: hl.eval(hl.nd.hstack([hl.nd.zeros((2, 0)), hl.nd.array([[1.0, 2.0], [3.0, 4.0]])]))
Out[3]:
array([[0., 2.],
       [0., 4.]])
```
The zeros matrix is 2 by 0, so horizontal concatenation should just
return the other matrix.
(I once saw the first column filled with random numbers, presumably from
a buffer overflow)

I did some cleaning up in the concat implementation, but the functional
change is to record the index of the first input which is non-empty
along the concat axis, and when resetting to the start of the axis,
reset to that non-empty index. Other size 0 inputs are correctly handled
when incrementing the index, the problem was that the first read happens
before an increment.