[query] if worker or driver raise an exception, fail the Batch job (#13715)
CHANGELOG: Fixes #13697, a long standing issue with QoB, in which a
failing partition job or driver job is not failed in the Batch UI.

I am not sure why we did not do this this way in the first place. If a
JVMJob raises an exception, Batch will mark the job as failed. Ergo, we
should raise an exception when a driver or a worker fails!

Here's an example: I used a simple pipeline that write to a bucket to
which I have read-only access. You can see an example Batch (where every
partition fails): https://batch.hail.is/batches/8046901. [1]

```python3
import hail as hl
hl.utils.range_table(3, n_partitions=3).write('gs://neale-bge/foo.ht')
```

NB: I removed the `log.error` in `handleForPython` because that log is
never necessary. That function converts a stack of exceptions into a
triplet of the short message, the full exception with stack trace, and a
Hail error id (if present). That triplet is always passed along to
someone else who logs the exception.

(FWIW, the error id indicates a Python source location that is
associated with the error. On the Python-side, we can look up that error
id and provide a better stack trace.)

[1] You'll notice the logs are missing. I noticed this as well, it's a
new bug. I fixed it in https://github.com/hail-is/hail/pull/13729.