[fs] improve copy tool progress bars
I renamed RichProgressBar and SimpleRichProgressBar to ...CopyToolProgressBar because
that is more accurate. I enhanced both to now include a count and a rate with the
right units based on the description. It is a bit flaky because I need the descriptions
to be exactly "files" or exactly "bytes" to pick the right units, but
this seems fine for the specific case of th CopyToolProgressBar.

There is probably a better way to build these UIs. I am sure we will
start to figure that out as we use rich more.