[query] faster array decoder (#13787)
Picking up where #13776 left off.

CHANGELOG: improved speed of reading hail format datasets from disk

This PR speeds up decoding arrays in two main ways:
* instead of calling `arrayType.isElementDefined(array, i)` on every
single array element, which expands to
  ```scala
  val b = aoff + lengthHeaderBytes + (i >> 3)
  !((Memory.loadByte(b) & (1 << (i & 7).toInt)) != 0)
  ```
process elements in groups of 64, and load the corresponding long of
missing bits once
* once we have a whole long of missing bits, we can be smarter than
branching on each bit. After flipping to get `presentBits`, we use the
following psuedocode to extract the positions of the set bits, with time
proportional to the number of set bits:
  ```
  while (presentBits != 0) {
    val idx = java.lang.Long.numberOfTrailingZeroes(presentBits)
    // do something with idx
presentBits = presentBits & (presentBits - 1) // unsets the rightmost
set bit
  }
  ```

To avoid needing to handle the last block of 64 elements differently,
this PR changes the layout of `PCanonicalArray` to ensure the missing
bits are always padded out to a multiple of 64 bits. They were already
padded to a multiple of 32, and I don't expect this change to have much
of an effect. But if needed, blocking by 32 elements instead had very
similar performance in my benchmarks.

I also experimented with unrolling loops. In the non-missing case, this
is easy. In the missing case, I tried using `if (presentBits.bitCount >=
8)` to guard an unrolled inner loop. In both cases, unrolling was if
anything slower.

Dan observed benefit from unrolling, but that was combined with the
first optimization above (not loading a bit from memory every element),
which I beleive was the real source of improvement.