[batch] Setup database infrastructure for job groups (#13475)
The conceptual change here is we want to parameterize all batch related
tables to have a new job group ID that I've set to **0** for the root
job group. We need to make sure all future inserts / updates into the
batches table are propagated to the new job groups table. When we create
a batch now, we also create the corresponding entries into the job
groups and job group parents tables.

I chose the root job group to be 0 as I think conceptually, the client
should start numbering job groups at 1 and not know there is a hidden
root job group being created under the hood. I'm not wedded to this.

I tried to check for all the indices that would be needed in my
prototype. It's possible I missed one or two, but it's not a big deal to
add it later.

I don't think we need to test this on a populated database (dev deploy
main, submit jobs, then run the migration), but let me know if you think
that would be helpful.