[fs] rate limit our use of parallelism in copy (#13833)
This is the result of some experimentation. With ten-way parallelism,
the copier very rarely gets rate-limited. With 75-way parallelism (the
default), we almost always experience a tens of transient errors. If we
start at ten and back off as in this PR, I can get to 75 with just a
handful of transient errors.

cc: @jigold