[hailtop] types, new exceptions, better task exception handling (#13818)
Besides the types and some transient exceptions, I think I fixed our
task exception handling in several spots. Two things:

1. We do not need to wait on a cancelled task. If it was not done, then
it could not possibly have an exception to retrieve. Moreover, now that
it is cancelled, there is nothing else to do. Cancellation is immediate.

2. If a task is done, we *must* always retrieve the exception, otherwise
we might not see an exception.