[benchmark] Benchmarks on Ubuntu 22.04 (#13880)
benchmark make targets are now made directly through the make targets in
`benchmark/Makefile`. Previous targets in `hail/Makefile` have been
deleted.
To submit benchmarks in batch from the root directory, use
```
$ BENCHMARK_LABEL=label NAMESPACE=default BENCHMARK_PROJECT=your-google-project make -C benchmark submit
```
The benchmark docker image is now based of `hail-ubuntu` with a minimal
set of requirements installed on top.
The hail wheel is added after the benchmark wheel in attempt to optimise
build times for changes to hail code.