[fs] improve copy tool progress bars (#13832)
I renamed RichProgressBar and SimpleRichProgressBar to
...CopyToolProgressBar because that is more accurate. I enhanced both to
now include a count and a rate with the right units based on the
description. It is a bit flaky because I need the descriptions to be
exactly "files" or exactly "bytes" to pick the right units, but this
seems fine for the specific case of th CopyToolProgressBar.

There is probably a better way to build these UIs. I am sure we will
start to figure that out as we use rich more.


Before:
<img width="830" alt="Screenshot 2023-10-16 at 18 28 14"
src="https://github.com/hail-is/hail/assets/106194/95f8828e-beb3-46d2-9403-18ff7aa60256">

After:
<img width="830" alt="Screenshot 2023-10-16 at 18 27 53"
src="https://github.com/hail-is/hail/assets/106194/01186b7c-d59f-4a0e-a1f6-9279fb50ae7e">