[devbin] Local UI development proxy server (#13632)
This adds a script that runs an `aiohttp` server on localhost that by
default proxies all requests to whatever batch namespace `hailctl` is
pointed to but serves static assets and renders all HTML locally. This
allows for local editing of HTML / CSS / SCSS files.

To run, this branch needs to be deployed in the relevant namespace (to
pick up the changes to `web_common.py`), and you need to have the `sass`
CLI installed (`brew install sass/sass/sass`). Then run in a separate
terminal `make devserver SERVICE=batch` and open up `localhost:8000`. If
you want to supply fake data instead of proxying an endpoint to the k8s
deployment of the service, you can override that endpoint in
`dev_proxy.py`.

If this looks good, I'll add the above instructions to the dev docs.

Resolves #13629