[query] fix bug in new dict decoder (#13939)
The decoder uses a `StagedArrayBuilder` to hold elements while being
sorted. The array builder is stored in a class field. When the same
decoder function is called more than once, that array builder is reused.

Before this fix, the array builder was never cleared, so if the decoder
function was called more than once, the array builder would still
contain the elements from previously decoded dicts.

Since it's highly non-obvious that you need to call `clear` immediately
after `new StagedCodeBuilder`, this PR makes the constructor take a
CodeBuilder, and always inserts a clear at the call site. I also took
the opportunity to CodeBuilderify the rest of the interface.